LB-20 refac: review fixes
